pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentbench"
version = "0.1.0"
description = "Pytest for AI agents. Define scenarios in YAML, run any agent, get results."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [{ name = "Daniel López Orta", email = "daniellopezorta39@gmail.com" }]
keywords = ["agent", "benchmark", "evaluation", "llm", "testing", "langgraph", "crewai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.optional-dependencies]
llm = ["openai>=1.0", "tiktoken>=0.7.0"]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "ruff>=0.4", "mypy>=1.10"]
all = ["openai>=1.0", "tiktoken>=0.7.0"]

[project.scripts]
agentbench = "agentbench.cli:main"

[project.urls]
Homepage = "https://github.com/daniellopez882/agentbench"
Repository = "https://github.com/daniellopez882/agentbench"
Issues = "https://github.com/daniellopez882/agentbench/issues"

[tool.ruff]
target-version = "py310"
line-length = 88

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
